fix: make test compile on Windows by disabling the subprocess runner
