+++ router-tests/runner_windows_test.go
@@ -1,0 +1,9 @@
+//go:build windows
+
+package integration_test
+
+import "errors"
+
+func NewSubprocessSubgraphsRunner() (SubgraphsRunner, error) {
+	return nil, errors.New("this subgraphs runner is not supported on Windows, use in-process or external instead")
+}
